userCustomerJWT

+++ src/main/java/com/diplom/restoran/repository/AdminRepository.java
@@ -7,5 +7,6 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AdminRepository extends JpaRepository<Admin, Integer>, PagingAndSortingRepository<Admin, Integer> {
+public interface AdminRepository extends JpaRepository<Admin, Long>, PagingAndSortingRepository<Admin, Long> {
+
 }
