+++ src/main/java/com/diplom/restoran/repository/WaiterRepository.java
@@ -8,6 +8,6 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface WaiterRepository extends JpaRepository<Waiter, Integer>, PagingAndSortingRepository<Waiter, Integer> {
+public interface WaiterRepository extends JpaRepository<Waiter, Long>, PagingAndSortingRepository<Waiter, Long> {
     public Optional<Waiter> findById (Long id);
 }
